Added more logic to help with round robin-ing the load
You can pass in two extra variables:

Cyclic Offset
Total Consumers

What it does it makes every consumer wait X number of milliseconds
before attempting to place a lock on the job. If there are 4 consumers,
and we are process 0, with an offset of 0, we will wait a rolling number
of milliseconds like so:

0, 1, 2, 3, 0, 1, 2, 3, 0, 1, 2, 3

This is to attempt to make the load more balanced between all workers as
the first 2 or 3 workers that start listening usually grab most of the
work that comes in.

Cyclic offset tell the process where to start in the delays. If you have
4 processes, then you use 0, 1, 2, 3 across all of them.